@@ -23,7 +23,7 @@ import android.content.Intent;
 import android.net.nsd.NsdServiceInfo;
 import android.net.nsd.NsdManager;
 import android.os.Build;
-import android.support.v4.content.LocalBroadcastManager;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 import android.util.Log;
 import android.widget.Toast;
 
