@@ -1,6 +1,6 @@
 package uk.ac.standrews.cs.mamoc_client.ServiceDiscovery;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
