@@ -2,7 +2,7 @@ package uk.ac.standrews.cs.mamoc_client.ServiceDiscovery;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.v4.content.LocalBroadcastManager;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 import android.util.Log;
 
 import java.util.TreeSet;
