@@ -2,7 +2,7 @@ package uk.ac.standrews.cs.mamoc_client.Utils;
 
 import android.app.Activity;
 import android.content.DialogInterface;
-import android.support.v7.app.AlertDialog;
+import androidx.appcompat.app.AlertDialog;
 import android.util.Log;
 import android.widget.Toast;
 
